requirements.txt
asgiref==3.7.2
blinker==1.6.2
boto3==1.26.127
botocore==1.29.127
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.1
Flask==2.3.2
idna==3.4
image==1.5.33
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
Pillow==9.5.0
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
Werkzeug==2.3.4
